Get rid of AdditionalValidators and code duplication. Contains breaking changes! (#404)
* Fix for large database

* Update sql queries

* Remove AdditionalValidators

* Up version because of breaking changes

+++ pg-index-health-model/src/main/java/io/github/mfvanek/pg/model/BloatAware.java
@@ -29,5 +29,5 @@ public interface BloatAware {
      *
      * @return bloat percentage
      */
-    int getBloatPercentage();
+    double getBloatPercentage();
 }
